Code refactoring based on SonarLint

+++ src/pl/pwr/hiervis/util/ui/HuePicker.java
@@ -14,15 +14,13 @@ import javax.swing.SwingUtilities;
 import pl.pwr.hiervis.util.HSV;
 import pl.pwr.hiervis.util.Utils;
 
-
 /**
  * A component which allows the user to select a color's hue.
  * 
  * @author Tomasz Bachmiński
  *
  */
-public class HuePicker extends JComponent
-{
+public class HuePicker extends JComponent {
 	public static final String SELECTED_HUE = HuePicker.class.getCanonicalName() + ":SELECTED_HUE";
 
 	private static final long serialVersionUID = -614405052134068408L;
@@ -37,179 +35,156 @@ public class HuePicker extends JComponent
 	private boolean paintHuesMaximized = false;
 	private boolean paintHueBorder = true;
 
-
-	public HuePicker()
-	{
+	public HuePicker() {
 		super();
 
 		MouseAdapter ml = new MouseAdapter() {
 			@Override
-			public void mousePressed( MouseEvent e )
-			{
-				setSelection( e );
+			public void mousePressed(MouseEvent e) {
+				setSelection(e);
 			}
 
 			@Override
-			public void mouseDragged( MouseEvent e )
-			{
-				setSelection( e );
+			public void mouseDragged(MouseEvent e) {
+				setSelection(e);
 			}
 		};
 
-		this.addMouseListener( ml );
-		this.addMouseMotionListener( ml );
+		this.addMouseListener(ml);
+		this.addMouseMotionListener(ml);
 	}
 
-	public void setIndicatorSize( int size )
-	{
-		if ( size <= 0 )
-			throw new IllegalArgumentException( "Size must be greater than 0." );
+	public void setIndicatorSize(int size) {
+		if (size <= 0)
+			throw new IllegalArgumentException("Size must be greater than 0.");
 		indicatorSize = size;
 		repaint();
 	}
 
-	public int getIndicatorSize()
-	{
+	public int getIndicatorSize() {
 		return indicatorSize;
 	}
 
 	/**
-	 * If set to true, the hue band will be painted to occupy the entire width
-	 * of this component, without trying to fit it in between the selection
-	 * indicators.
-	 * This is useful for when the component has very small width compared
-	 * to the selection indicators.
+	 * If set to true, the hue band will be painted to occupy the entire width of
+	 * this component, without trying to fit it in between the selection indicators.
+	 * This is useful for when the component has very small width compared to the
+	 * selection indicators.
 	 */
-	public void setPaintHuesMaximized( boolean value )
-	{
+	public void setPaintHuesMaximized(boolean value) {
 		paintHuesMaximized = value;
 		repaint();
 	}
 
-	public boolean getPaintHuesMaximized()
-	{
+	public boolean getPaintHuesMaximized() {
 		return paintHuesMaximized;
 	}
 
 	/**
 	 * If set to true, the hue band will be painted with a 1px border around it.
 	 */
-	public void setPaintHueBorder( boolean value )
-	{
+	public void setPaintHueBorder(boolean value) {
 		paintHueBorder = value;
 		repaint();
 	}
 
-	public boolean getPaintHueBorder()
-	{
+	public boolean getPaintHueBorder() {
 		return paintHueBorder;
 	}
 
-	public void setSelection( float hue )
-	{
-		if ( hue < 0 || hue > 1.0f )
-			throw new IllegalArgumentException( "0 < " + hue + " < 1.0" );
+	public void setSelection(float hue) {
+		if (hue < 0 || hue > 1.0f)
+			throw new IllegalArgumentException("0 < " + hue + " < 1.0");
 		float o = selectedHue;
 		selectedHue = hue;
 		repaint();
-		firePropertyChange( SELECTED_HUE, o, hue );
+		firePropertyChange(SELECTED_HUE, o, hue);
 	}
 
-	public void setSelection( Color color )
-	{
-		setSelection( new HSV( color ).getHue() );
+	public void setSelection(Color color) {
+		setSelection(new HSV(color).getHue());
 	}
 
-	public float getSelection()
-	{
+	public float getSelection() {
 		return selectedHue;
 	}
 
-	private void setSelection( MouseEvent e )
-	{
-		if ( SwingUtilities.isLeftMouseButton( e ) ) {
+	private void setSelection(MouseEvent e) {
+		if (SwingUtilities.isLeftMouseButton(e)) {
 			final int minY = indicatorSize / 2;
 			final int maxY = getHeight() - indicatorSize / 2;
 
-			float y = Utils.clamp( minY, e.getY(), maxY ) - minY;
-			setSelection( 1f - ( y / ( maxY - minY ) ) );
+			float y = Utils.clamp(minY, e.getY(), maxY) - (float) minY;
+			setSelection(1f - (y / (maxY - minY)));
 		}
 	}
 
 	@Override
-	public void paintComponent( Graphics g )
-	{
-		super.paintComponent( g );
+	public void paintComponent(Graphics g) {
+		super.paintComponent(g);
 
-		Graphics2D g2 = (Graphics2D)g;
-		g2.setRenderingHint(
-			RenderingHints.KEY_ANTIALIASING,
-			RenderingHints.VALUE_ANTIALIAS_ON
-		);
+		Graphics2D g2 = (Graphics2D) g;
+		g2.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON);
 
-		paintHues( g2 );
-		paintSelectionIndicator( g2 );
+		paintHues(g2);
+		paintSelectionIndicator(g2);
 	}
 
-	protected void paintHues( Graphics g )
-	{
+	protected void paintHues(Graphics g) {
 		final int x;
 		final int w;
 		final int h = getHeight() - indicatorSize;
 		final float stepY = 1f / h;
 
-		if ( paintHuesMaximized ) {
-			if ( paintHueBorder ) {
+		if (paintHuesMaximized) {
+			if (paintHueBorder) {
 				x = 1;
 				w = getWidth() - 2;
-			}
-			else {
+			} else {
 				x = 0;
 				w = getWidth();
 			}
-		}
-		else {
+		} else {
 			x = indicatorSize;
 			w = getWidth() - 2 * indicatorSize;
 		}
 
-		if ( paintHueBorder ) {
-			g.setColor( hueBorderColor );
-			g.fillRect( x - 1, indicatorSize / 2 - 1, w + 2, h + 2 );
+		if (paintHueBorder) {
+			g.setColor(hueBorderColor);
+			g.fillRect(x - 1, indicatorSize / 2 - 1, w + 2, h + 2);
 		}
 
-		for ( int y = 0; y < h; ++y ) {
-			Color c = new Color( Color.HSBtoRGB( 1f - y * stepY, 1, 1 ) );
-			g.setColor( c );
-			g.fillRect( x, y + indicatorSize / 2, w, 1 );
+		for (int y = 0; y < h; ++y) {
+			Color c = new Color(Color.HSBtoRGB(1f - y * stepY, 1, 1));
+			g.setColor(c);
+			g.fillRect(x, y + indicatorSize / 2, w, 1);
 		}
 	}
 
-	protected void paintSelectionIndicator( Graphics g )
-	{
+	protected void paintSelectionIndicator(Graphics g) {
 		final int w = getWidth();
 		final int h = getHeight() - indicatorSize;
 
-		int hOffset = Math.round( ( 1f - selectedHue ) * h );
-		if ( hOffset == h )
+		int hOffset = Math.round((1f - selectedHue) * h);
+		if (hOffset == h)
 			hOffset -= 1;
 		hOffset += indicatorSize / 2;
 
 		Polygon leftTri = new Polygon();
-		leftTri.addPoint( 0, hOffset - indicatorSize / 2 );
-		leftTri.addPoint( 0, hOffset + indicatorSize / 2 );
-		leftTri.addPoint( indicatorSize, hOffset );
+		leftTri.addPoint(0, hOffset - indicatorSize / 2);
+		leftTri.addPoint(0, hOffset + indicatorSize / 2);
+		leftTri.addPoint(indicatorSize, hOffset);
 
 		Polygon rightTri = new Polygon();
-		rightTri.addPoint( w - 1, hOffset - indicatorSize / 2 );
-		rightTri.addPoint( w - 1, hOffset + indicatorSize / 2 );
-		rightTri.addPoint( w - indicatorSize - 1, hOffset );
+		rightTri.addPoint(w - 1, hOffset - indicatorSize / 2);
+		rightTri.addPoint(w - 1, hOffset + indicatorSize / 2);
+		rightTri.addPoint(w - indicatorSize - 1, hOffset);
 
-		g.setColor( selectionFillColor );
-		g.fillPolygon( leftTri );
-		g.fillPolygon( rightTri );
-		g.setColor( selectionBorderColor );
-		g.drawPolygon( leftTri );
-		g.drawPolygon( rightTri );
+		g.setColor(selectionFillColor);
+		g.fillPolygon(leftTri);
+		g.fillPolygon(rightTri);
+		g.setColor(selectionBorderColor);
+		g.drawPolygon(leftTri);
+		g.drawPolygon(rightTri);
 	}
 }
